fix : try to fix error file unexpected in json and use DP to menu both of them
