+++ HenriksHobbyLager/appsettings.json
@@ -1,0 +1,10 @@
+{
+  "ConnectionStrings": {
+    "SqlConnection": "Server=localhost,1433;Database=YourDatabaseName;User Id=YourUsername;Password=YourPassword;TrustServerCertificate=True",
+    "MongoConnection": "mongodb+srv://username:password@cluster.mongodb.net/myDatabase"
+  },
+  "DatabaseSettings": {
+    "SqlType": "SQL",
+    "NoSqlType": "NoSQL"
+  }
+}
